Updates fixture for store
